requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2>=2.8
Pillow>=8.4.0
pipreqs>=0.4.11
django-simple-history>=3.0.0
boto3>=1.20.24
django-storages>=1.12.3
django-cors-headers>=3.10.1